+++ src/components/community/write/CommunityWrite.tsx
@@ -3,8 +3,8 @@ import * as W from './WriteStyle';
 import { Back } from '../detail/DetailStyle';
 import { useState, useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import request from '../../../utils/request';
-import { axiosInstance } from '../../../utils/axios';
+import request from '../../../api/request';
+import { axiosInstance } from '../../../api/axios';
 import ImageUpload from '../../utils/ImageUpload';
 import { ReactComponent as ArrowIcon } from '../../../img/community/arrow_left.svg';
 import { useAuth } from '../../../hooks/useAuth';
