+++ src/query/get/useGetComment.tsx
@@ -1,5 +1,5 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
-import request from '../../utils/request';
+import request from '../../api/request';
 import { PostComment } from '../../components/community/detail/CommentData';
 
 interface useGetCommentProps {
