+++ src/query/get/useGetPostDetail.tsx
@@ -1,5 +1,5 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
-import request from '../../utils/request';
+import request from '../../api/request';
 import { Post } from '../../components/community/detail/CommentData';
 
 interface useGetDonatePostProps {
